blog/requirements.txt
Django==3.1.2
djangorestframework==3.12.1
Pillow==8.0.0
django-cors-headers==3.5.0